Split index.d.ts add types for aws-lambda
