+++ package.json
@@ -19,6 +19,7 @@
     },
     "dependencies": {
         "ajv": "8.10.0",
+        "aws-lambda": "^1.0.7",
         "aws-sdk": "2.1105.0",
         "js-yaml": "4.1.0",
         "json-schema-merge-allof": "^0.8.1",
@@ -26,6 +27,7 @@
         "xml2js": "0.4.23"
     },
     "devDependencies": {
+        "@types/aws-lambda": "^8.10.93",
         "chai": "^4.2.0",
         "chai-as-promised": "^7.1.1",
         "eslint": "^7.7.0",
