+++ typings/require-at-least-one.d.ts
